Moves Pratt parser tests into separate file
Moves the example showcasing Pratt parser features into a separate file.
