+++ Examples/LeftVsRightRecursion.test.ts
@@ -235,79 +235,3 @@ describe("Left Recursive Pratt", () => {
         );
     });
 });
-
-describe("Left Recursive Pratt With Pre and Postfix", () => {
-    const left = int();
-    const infix = oneOf(
-        toBinaryOperator(exact("-"), [1, 2]),
-        toBinaryOperator(exact("/"), [3, 4]),
-    );
-    const prefix = toUnaryOperator(exact("+"), 5);
-    const postfix = toUnaryOperator(exact("!"), 7);
-    const scopeBegin = exact("{");
-    const scopeEnd = exact("}");
-
-    describe("as S Expressions", () => {
-        const exprParser = pratt<number, string>(left, {
-            infix: {
-                op: infix,
-                map: (symbol, left, right) => `(${symbol} ${left} ${right})`,
-            },
-            prefix: {
-                op: prefix,
-                map: (symbol, right) => `(${symbol} ${right})`,
-            },
-            postfix: {
-                op: postfix,
-                map: (symbol, left) => `(${symbol} ${left})`,
-            },
-            scope: {
-                scopeBegin,
-                scopeEnd,
-            },
-        });
-
-        const okCases: Array<[string, string | number]> = [
-            ["1 - 2 - 3 - 4", "(- (- (- 1 2) 3) 4)"],
-            ["1 / 2 / 3 / 4", "(/ (/ (/ 1 2) 3) 4)"],
-            ["+3", "(+ 3)"],
-            ["3!", "(! 3)"],
-            ["++3", "(+ (+ 3))"],
-            ["3!!", "(! (! 3))"],
-            ["3!!!!", "(! (! (! (! 3))))"],
-            ["++++3", "(+ (+ (+ (+ 3))))"],
-            ["+3!", "(+ (! 3))"], // "!" has higher binding power / precedence than "+"
-            ["++3!!", "(+ (+ (! (! 3))))"],
-            ["1 / 2 / +3", "(/ (/ 1 2) (+ 3))"],
-            ["1 / 2 / +3 / 4", "(/ (/ (/ 1 2) (+ 3)) 4)"],
-            ["1 / 2 / +3 / + 4", "(/ (/ (/ 1 2) (+ 3)) (+ 4))"],
-            ["1 / 2 / 3! / + 4", "(/ (/ (/ 1 2) (! 3)) (+ 4))"],
-            ["1 / 2 / +3! / + 4", "(/ (/ (/ 1 2) (+ (! 3))) (+ 4))"],
-            ["1", 1],
-            ["{1}", 1],
-            ["{+1}", "(+ 1)"],
-            ["{1!}", "(! 1)"],
-            ["{+1}!", "(! (+ 1))"],
-            ["+{+3}!!", "(+ (! (! (+ 3))))"],
-            ["{1 / 2}", "(/ 1 2)"],
-            ["1 / {2 / 3}", "(/ 1 (/ 2 3))"],
-            ["1 / {2 / 3} / 4", "(/ (/ 1 (/ 2 3)) 4)"],
-            ["1 / {2 / 3 / 4}", "(/ 1 (/ (/ 2 3) 4))"],
-        ];
-
-        test.each(okCases)(
-            "parses '%s' as the string '%s'",
-            (source, expectedValue) => {
-                let result = exprParser.parse(source);
-                switch (result.variant) {
-                    case Result.Variant.Ok:
-                        expect(result.value[0]).toBe(expectedValue);
-                        break;
-
-                    case Result.Variant.Err:
-                        fail(result.error);
-                }
-            },
-        );
-    });
-});
